Propagation of data and socket creation separated
