+++ src/index.jsx
@@ -7,6 +7,7 @@ import { WorldMap } from './components/WorldMap';
 import connect from './connect';
 
 const events = connect('http://127.0.0.1:8090');
+
 const WorldMapContainer = withStore(events, 'players')(WorldMap);
 
 ReactDOM.render(
